(Major) Journey engine started
Started to work on a prequel to the Odyssey engine called Journey. Essentially stripping the extra features I added for Odyssey that were built on Journey anyways (before Journey had a name it was really just a few helpful sub-systems.

Journey will be a stripped down version of Odyssey that's less opinionated and gives you the freedom to essentially plug-and-play by adding the sprites to an already existing project or allowing you to start fresh with a whole set of nice features.

Journey will be a lightweight game engine technically, but it's more like heavy framework as it won't have any opinions of how you should structure code, physics, or anything like that. It will only give you tools to work with and is made to work 100% with vanilla Scratch (and it's workflow!).

Yes most projects will need to do some work to integrate into the broadcasts and tools Journey gives you, but existing code might not necessary need to be migrated and there will be resources to help with the integration anyways.