Feature: Basic save/load system
Added a working system for saving and loading that will be improved on to allow for any number of scenes to be loaded with any amount of objects. Currently only supports 1 scene as scenes aren't their own asset yet technically and the thing that is being loaded is a save file (which also contains all state of the scene). Kind of a mess right now but it's just a working prototype basically.

- Added save system (replaced save manager)
- Added a complete set of YAML functions to load the data from the deserialized YAML string
- Added a way to pass in any string to be deserialized
- Added a load method for objects to load save data in a structured way. Counts as a lifecycle but not properly integrated as one (since it's state doesn't need to be saved unlike other lifecycles)
- Improved finding keys and loading data
- No longer clears the loaded YAML when stopping the project
- Maybe more changes I forgot